funk: add placeholder accounts_lru_prev_idx/accounts_lru_next_idx rec fields

+++ src/flamenco/runtime/tests/harness/fd_harness_common.c
@@ -8,7 +8,7 @@ fd_runtime_fuzz_runner_align( void ) {
 ulong
 fd_runtime_fuzz_runner_footprint( void ) {
   ulong txn_max = 4+fd_tile_cnt();
-  ulong rec_max = 1024UL;
+  uint rec_max  = 1024;
 
   ulong l = FD_LAYOUT_INIT;
   l = FD_LAYOUT_APPEND( l, fd_runtime_fuzz_runner_align(), sizeof(fd_runtime_fuzz_runner_t) );
@@ -22,7 +22,7 @@ fd_runtime_fuzz_runner_new( void * mem,
                             void * spad_mem,
                             ulong  wksp_tag ) {
   ulong txn_max = 4+fd_tile_cnt();
-  ulong rec_max = 1024UL;
+  uint rec_max  = 1024;
 
   FD_SCRATCH_ALLOC_INIT( l, mem );
   void * runner_mem = FD_SCRATCH_ALLOC_APPEND( l, fd_runtime_fuzz_runner_align(), sizeof(fd_runtime_fuzz_runner_t) );
